PHRAS-3442_optimize-list-notifications_4.1-bis WIP : pushed to run all tests on circle move "session/notifications" controller to "user/notifications" use twig to render notifs + dropdown + dlg fixed some tests still todo : mark "read"
