--- templates/web/prod/notifications.html.twig
@@ -1,31 +1,0 @@
-<div style="margin-right:16px;">
-
-{% if notifications|length == 0%}
-<div class="notification_title">
-  <span>{{ 'Aucune notification' | trans }}</span>
-</div>
-{% else %}
-<div class="notification_title">
-  <a href="#" class="notification__print-action">{{ 'toutes les notifications' | trans }}</a>
-</div>
-{% endif %}
-
-{% for notif in notifications %}
-<div style="position:relative;" id="notification_{{notif['id']}}" class="notification {% if notif['unread'] == '1' %}unread{% endif %}">
-  <table style="width:100%;" cellspacing="0" cellpadding="0" border="0">
-    <tr>
-      <td style="width:25px;">
-        <img src="{{notif['icon']}}" style="vertical-align:middle;width:16px;margin:2px;" />
-      </td>
-      <td>
-        <p style="margin:0;padding:0;" class="{{notif['class']}}">
-          {{notif['text']|raw}}
-          <span class="time">{{notif['created_on']}}</span>
-        </p>
-      </td>
-    </tr>
-  </table>
-</div>
-
-{% endfor %}
-</div>
