+++ templates/web/prod/notifications_dialog.html.twig
@@ -1,0 +1,7 @@
+<div id="notifications-dialog">
+    <div class="notifications">
+    </div>
+    <div class="navigation">
+        <a href="#" class="notification__print-action" data-offset="0">{{ 'charger d\'avantages de notifications' }}</a>
+    </div>
+</div>
