+++ templates/web/prod/notifications_dropdown.html.twig
@@ -1,0 +1,12 @@
+<div style="display:none;z-index:30000;" id="notification_box">
+    <div style="margin-right:16px;">
+        <div class="no_notifications">
+            <span>{{ 'Aucune notification' | trans }}</span>
+        </div>
+        <div class="notifications">
+        </div>
+        <div class="show_all">
+            <a href="#" class="notification__print-action">{{ 'toutes les notifications' | trans }}</a>
+        </div>
+    </div>
+</div>
